+++ src/api/fetcher.ts
@@ -1,3 +1,5 @@
+import { useUserPersistStore } from '@/stores/userPersistStore';
+
 export async function fetcher<T>(
   url: string,
   options?: RequestInit,
@@ -13,6 +15,10 @@ export async function fetcher<T>(
       credentials: 'include',
       signal: AbortSignal.timeout(5000)
     });
+
+  if (res.status === 401) {
+    useUserPersistStore.getState().cleanUser();
+  }
 
   if (!res.ok) {
     throw new Error(`Fetch error: ${res.status}`);
